Fix non-functional MTP.
This requires a minor patch to the kernel:

  https://github.com/eldracode/kernel_samsung_m21/commit/b2ccd97fb90f179ea0c7d273b4b053811e282765

and a kernel config change to avoid a reproducible spontaneous reboot
after 220 seconds of uptime in TWRP:

  https://github.com/eldracode/kernel_samsung_m21/commit/59d86f38dcc3936b2ffae9a9f7fe8fc82adc98ab

The default TWRP init.rc is now used in favour of Samsung's recovery
init.rc, and a custom USB configuration is applied.

Additionally, MTP no longer needs to be disabled to allow adb to work.

Signed-off-by: Ayush <ayush@disroot.org>